+++ fe/fe-core/src/test/java/org/apache/doris/datasource/DatasourceMgrTest.java
@@ -20,12 +20,21 @@ package org.apache.doris.datasource;
 import org.apache.doris.analysis.AlterCatalogNameStmt;
 import org.apache.doris.analysis.AlterCatalogPropertyStmt;
 import org.apache.doris.analysis.CreateCatalogStmt;
+import org.apache.doris.analysis.CreateRoleStmt;
+import org.apache.doris.analysis.CreateUserStmt;
 import org.apache.doris.analysis.DropCatalogStmt;
+import org.apache.doris.analysis.GrantStmt;
 import org.apache.doris.analysis.ShowCatalogStmt;
+import org.apache.doris.analysis.SwitchStmt;
+import org.apache.doris.analysis.UserIdentity;
 import org.apache.doris.catalog.Catalog;
+import org.apache.doris.common.AnalysisException;
 import org.apache.doris.common.Config;
 import org.apache.doris.common.FeConstants;
+import org.apache.doris.mysql.privilege.PaloAuth;
+import org.apache.doris.qe.ConnectContext;
 import org.apache.doris.qe.ShowResultSet;
+import org.apache.doris.system.SystemInfoService;
 import org.apache.doris.utframe.TestWithFeService;
 
 import org.junit.Assert;
@@ -44,11 +53,60 @@ public class DatasourceMgrTest extends TestWithFeService {
     private DataSourceMgr mgr;
     private static final String MY_CATALOG = "my_catalog";
 
+    private static PaloAuth auth;
+    private static Catalog catalog;
+    private static UserIdentity user1;
+    private static UserIdentity user2;
+
     @Override
     protected void runBeforeAll() throws Exception {
         Config.enable_multi_catalog = true;
         FeConstants.runningUnitTest = true;
         mgr = Catalog.getCurrentCatalog().getDataSourceMgr();
+
+        ConnectContext rootCtx = createDefaultCtx();
+        catalog = Catalog.getCurrentCatalog();
+        auth = catalog.getAuth();
+
+        // grant with no catalog is switched, internal catalog works.
+        CreateRoleStmt createRole1 = (CreateRoleStmt) parseAndAnalyzeStmt("create role role1;", rootCtx);
+        auth.createRole(createRole1);
+        GrantStmt grantRole1 = (GrantStmt) parseAndAnalyzeStmt("grant grant_priv on tpch.* to role 'role1';", rootCtx);
+        auth.grant(grantRole1);
+        // grant with ctl.db.tbl. grant can succeed even if the catalog does not exist
+        GrantStmt grantRole1WithCtl = (GrantStmt) parseAndAnalyzeStmt(
+                "grant select_priv on testc.testdb.* to role 'role1';", rootCtx);
+        auth.grant(grantRole1WithCtl);
+        // user1 can't switch to hive
+        auth.createUser((CreateUserStmt) parseAndAnalyzeStmt(
+                "create user 'user1'@'%' identified by 'pwd1' default role 'role1';", rootCtx));
+        user1 = new UserIdentity("user1", "%");
+        user1.analyze(SystemInfoService.DEFAULT_CLUSTER);
+        // user1 has the privileges of testc which is granted by ctl.db.tbl format.
+        Assert.assertTrue(auth.getDbPrivTable().hasPrivsOfCatalog(user1, "testc"));
+
+        // create catalog
+        CreateCatalogStmt hiveCatalog = (CreateCatalogStmt) parseAndAnalyzeStmt(
+                "create catalog hive properties('type' = 'hms', 'hive.metastore.uris' = 'thrift://192.168.0.1:9083');",
+                rootCtx);
+        catalog.getDataSourceMgr().createCatalog(hiveCatalog);
+        CreateCatalogStmt iceBergCatalog = (CreateCatalogStmt) parseAndAnalyzeStmt(
+                "create catalog iceberg properties('type' = 'hms', 'iceberg.hive.metastore.uris' = 'thrift://192.168.0.1:9083');",
+                rootCtx);
+        catalog.getDataSourceMgr().createCatalog(iceBergCatalog);
+
+        // switch to hive.
+        SwitchStmt switchHive = (SwitchStmt) parseAndAnalyzeStmt("switch hive;", rootCtx);
+        catalog.changeCatalog(rootCtx, switchHive.getCatalogName());
+        CreateRoleStmt createRole2 = (CreateRoleStmt) parseAndAnalyzeStmt("create role role2;", rootCtx);
+        auth.createRole(createRole2);
+        GrantStmt grantRole2 = (GrantStmt) parseAndAnalyzeStmt("grant grant_priv on tpch.customer to role 'role2';",
+                rootCtx);
+        auth.grant(grantRole2);
+        auth.createUser((CreateUserStmt) parseAndAnalyzeStmt(
+                "create user 'user2'@'%' identified by 'pwd2' default role 'role2';", rootCtx));
+        user2 = new UserIdentity("user2", "%");
+        user2.analyze(SystemInfoService.DEFAULT_CLUSTER);
     }
 
     @Test
@@ -61,7 +119,7 @@ public class DatasourceMgrTest extends TestWithFeService {
         String showCatalogSql = "SHOW CATALOGS";
         ShowCatalogStmt showStmt = (ShowCatalogStmt) parseAndAnalyzeStmt(showCatalogSql);
         ShowResultSet showResultSet = mgr.showCatalogs(showStmt);
-        Assertions.assertEquals(2, showResultSet.getResultRows().size());
+        Assertions.assertEquals(4, showResultSet.getResultRows().size());
 
         String alterCatalogNameSql = "ALTER CATALOG hms_catalog RENAME " + MY_CATALOG + ";";
         AlterCatalogNameStmt alterNameStmt = (AlterCatalogNameStmt) parseAndAnalyzeStmt(alterCatalogNameSql);
@@ -71,14 +129,6 @@ public class DatasourceMgrTest extends TestWithFeService {
                 + " (\"type\" = \"hms\", \"k\" = \"v\");";
         AlterCatalogPropertyStmt alterPropStmt = (AlterCatalogPropertyStmt) parseAndAnalyzeStmt(alterCatalogProps);
         mgr.alterCatalogProps(alterPropStmt);
-
-        showResultSet = mgr.showCatalogs(showStmt);
-        for (List<String> row : showResultSet.getResultRows()) {
-            if (row.get(1).equals(InternalDataSource.INTERNAL_DS_NAME)) {
-                continue;
-            }
-            Assertions.assertEquals(MY_CATALOG, row.get(0));
-        }
 
         String showDetailCatalog = "SHOW CATALOG my_catalog";
         ShowCatalogStmt showDetailStmt = (ShowCatalogStmt) parseAndAnalyzeStmt(showDetailCatalog);
@@ -101,7 +151,7 @@ public class DatasourceMgrTest extends TestWithFeService {
         DropCatalogStmt dropCatalogStmt = (DropCatalogStmt) parseAndAnalyzeStmt(dropCatalogSql);
         mgr.dropCatalog(dropCatalogStmt);
         showResultSet = mgr.showCatalogs(showStmt);
-        Assertions.assertEquals(1, showResultSet.getResultRows().size());
+        Assertions.assertEquals(3, showResultSet.getResultRows().size());
     }
 
     private void testDataSourceMgrPersist() throws Exception {
@@ -123,7 +173,7 @@ public class DatasourceMgrTest extends TestWithFeService {
         DataInputStream dis = new DataInputStream(new FileInputStream(file));
         DataSourceMgr mgr2 = DataSourceMgr.read(dis);
 
-        Assert.assertEquals(2, mgr2.listCatalogs().size());
+        Assert.assertEquals(4, mgr2.listCatalogs().size());
         Assert.assertEquals(myCatalog.getId(), mgr2.getCatalog(MY_CATALOG).getId());
         Assert.assertEquals(0, mgr2.getInternalDataSource().getId());
         Assert.assertEquals(0, mgr2.getCatalog(InternalDataSource.INTERNAL_DS_ID).getId());
@@ -138,5 +188,69 @@ public class DatasourceMgrTest extends TestWithFeService {
         // 3. delete files
         dis.close();
         file.delete();
+    }
+
+    @Test
+    public void testSwitchCommand() throws Exception {
+        // mock the login of user1
+        ConnectContext user1Ctx = createCtx(user1, "127.0.0.1");
+        // user1 can switch to internal catalog
+        parseAndAnalyzeStmt("switch " + InternalDataSource.INTERNAL_DS_NAME + ";", user1Ctx);
+        Assert.assertEquals(InternalDataSource.INTERNAL_DS_NAME, user1Ctx.getDefaultCatalog());
+        // user1 can't switch to hive
+        try {
+            parseAndAnalyzeStmt("switch hive;", user1Ctx);
+            Assert.fail("user1 switch to hive with no privilege.");
+        } catch (AnalysisException e) {
+            Assert.assertEquals(e.getMessage(),
+                    "errCode = 2, detailMessage = Access denied for user 'default_cluster:user1' to catalog 'hive'");
+        }
+        Assert.assertEquals(InternalDataSource.INTERNAL_DS_NAME, user1Ctx.getDefaultCatalog());
+
+        // mock the login of user2
+        ConnectContext user2Ctx = createCtx(user2, "127.0.0.1");
+        // user2 can switch to internal catalog
+        parseAndAnalyzeStmt("switch " + InternalDataSource.INTERNAL_DS_NAME + ";", user2Ctx);
+        Assert.assertEquals(InternalDataSource.INTERNAL_DS_NAME, user2Ctx.getDefaultCatalog());
+        // user2 can switch to hive
+        SwitchStmt switchHive = (SwitchStmt) parseAndAnalyzeStmt("switch hive;", user2Ctx);
+        catalog.changeCatalog(user2Ctx, switchHive.getCatalogName());
+        Assert.assertEquals(user2Ctx.getDefaultCatalog(), "hive");
+        // user2 can grant select_priv to tpch.customer
+        GrantStmt user2GrantHiveTable = (GrantStmt) parseAndAnalyzeStmt(
+                "grant select_priv on tpch.customer to 'user2'@'%';", user2Ctx);
+        auth.grant(user2GrantHiveTable);
+    }
+
+    @Test
+    public void testShowCatalogStmtWithPrivileges() throws Exception {
+        // mock the login of user1
+        ConnectContext user1Ctx = createCtx(user1, "127.0.0.1");
+        ShowCatalogStmt user1Show = (ShowCatalogStmt) parseAndAnalyzeStmt("show catalogs;", user1Ctx);
+        List<List<String>> user1ShowResult = catalog.getDataSourceMgr().showCatalogs(user1Show).getResultRows();
+        Assert.assertEquals(user1ShowResult.size(), 1);
+        Assert.assertEquals(user1ShowResult.get(0).get(1), InternalDataSource.INTERNAL_DS_NAME);
+        Assert.assertEquals(user1ShowResult.get(0).get(0), String.valueOf(InternalDataSource.INTERNAL_DS_ID));
+
+        // mock the login of user2
+        ConnectContext user2Ctx = createCtx(user2, "127.0.0.1");
+        ShowCatalogStmt user2Show = (ShowCatalogStmt) parseAndAnalyzeStmt("show catalogs;", user2Ctx);
+        List<List<String>> user2ShowResult = catalog.getDataSourceMgr().showCatalogs(user2Show).getResultRows();
+        Assert.assertEquals(user2ShowResult.size(), 2);
+        Assert.assertTrue(user2ShowResult.stream().map(l -> l.get(1)).anyMatch(c -> c.equals("hive")));
+
+        // access denied
+        ShowCatalogStmt user2ShowHive = (ShowCatalogStmt) parseAndAnalyzeStmt("show catalog hive;", user2Ctx);
+        List<List<String>> user2ShowHiveResult = catalog.getDataSourceMgr().showCatalogs(user2ShowHive).getResultRows();
+        Assert.assertTrue(
+                user2ShowHiveResult.stream().map(l -> l.get(0)).anyMatch(c -> c.equals("hive.metastore.uris")));
+        try {
+            catalog.getDataSourceMgr()
+                    .showCatalogs((ShowCatalogStmt) parseAndAnalyzeStmt("show catalog iceberg;", user2Ctx));
+            Assert.fail("");
+        } catch (AnalysisException e) {
+            Assert.assertEquals(e.getMessage(),
+                    "errCode = 2, detailMessage = Access denied for user 'default_cluster:user2' to catalog 'iceberg'");
+        }
     }
 }
