+++ fe/fe-core/src/test/java/org/apache/doris/utframe/TestWithFeService.java
@@ -40,7 +40,6 @@ import org.apache.doris.common.Config;
 import org.apache.doris.common.DdlException;
 import org.apache.doris.common.FeConstants;
 import org.apache.doris.common.util.SqlParserUtils;
-import org.apache.doris.mysql.privilege.PaloAuth;
 import org.apache.doris.planner.Planner;
 import org.apache.doris.qe.ConnectContext;
 import org.apache.doris.qe.OriginStatement;
@@ -116,12 +115,16 @@ public abstract class TestWithFeService {
 
     // Help to create a mocked ConnectContext.
     protected ConnectContext createDefaultCtx() throws IOException {
+        return createCtx(UserIdentity.ROOT, "127.0.0.1");
+    }
+
+    protected ConnectContext createCtx(UserIdentity user, String host) throws IOException {
         SocketChannel channel = SocketChannel.open();
         ConnectContext ctx = new ConnectContext(channel);
         ctx.setCluster(SystemInfoService.DEFAULT_CLUSTER);
-        ctx.setCurrentUserIdentity(UserIdentity.ROOT);
-        ctx.setQualifiedUser(PaloAuth.ROOT_USER);
-        ctx.setRemoteIP("127.0.0.1");
+        ctx.setCurrentUserIdentity(user);
+        ctx.setQualifiedUser(user.getQualifiedUser());
+        ctx.setRemoteIP(host);
         ctx.setCatalog(Catalog.getCurrentCatalog());
         ctx.setThreadLocalInfo();
         return ctx;
@@ -129,11 +132,15 @@ public abstract class TestWithFeService {
 
     // Parse an origin stmt and analyze it. Return a StatementBase instance.
     protected StatementBase parseAndAnalyzeStmt(String originStmt) throws Exception {
+        return parseAndAnalyzeStmt(originStmt, connectContext);
+    }
+
+    // Parse an origin stmt and analyze it. Return a StatementBase instance.
+    protected StatementBase parseAndAnalyzeStmt(String originStmt, ConnectContext ctx) throws Exception {
         System.out.println("begin to parse stmt: " + originStmt);
-        SqlScanner input =
-                new SqlScanner(new StringReader(originStmt), connectContext.getSessionVariable().getSqlMode());
+        SqlScanner input = new SqlScanner(new StringReader(originStmt), ctx.getSessionVariable().getSqlMode());
         SqlParser parser = new SqlParser(input);
-        Analyzer analyzer = new Analyzer(connectContext.getCatalog(), connectContext);
+        Analyzer analyzer = new Analyzer(ctx.getCatalog(), ctx);
         StatementBase statementBase = null;
         try {
             statementBase = SqlParserUtils.getFirstStmt(parser);
@@ -154,8 +161,8 @@ public abstract class TestWithFeService {
     // for analyzing multi statements
     protected List<StatementBase> parseAndAnalyzeStmts(String originStmt) throws Exception {
         System.out.println("begin to parse stmts: " + originStmt);
-        SqlScanner input =
-                new SqlScanner(new StringReader(originStmt), connectContext.getSessionVariable().getSqlMode());
+        SqlScanner input = new SqlScanner(new StringReader(originStmt),
+                connectContext.getSessionVariable().getSqlMode());
         SqlParser parser = new SqlParser(input);
         Analyzer analyzer = new Analyzer(connectContext.getCatalog(), connectContext);
         List<StatementBase> statementBases = null;
